Monster module can now return connection type. Renamed BaseModel into BaseType
