+++ src/schema/monsters/monster.args.ts
@@ -1,10 +1,5 @@
-import { Length } from 'class-validator';
-import { ArgsType, Field } from 'type-graphql';
+import { ArgsType } from 'type-graphql';
 import { PaginationArgs } from '../shared';
 
 @ArgsType()
-export class MonsterArgs extends PaginationArgs {
-  @Field({ nullable: true })
-  @Length(1)
-  name?: string;
-}
+export class MonsterArgs extends PaginationArgs {}
