+++ src/schema/shared/baseType.ts
@@ -1,0 +1,7 @@
+import { Field, ID, InterfaceType } from 'type-graphql';
+
+@InterfaceType()
+export abstract class BaseType {
+  @Field(() => ID)
+  readonly id: number;
+}
